requirements.txt
PyQt6>=6.4.0
pyyaml>=6.0
requests>=2.28.0
inotify>=0.9.6
